# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CheckPasswordRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  dto.DuplicateEmailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      email:
        example: user@example.com
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsIn...
        type: string
      user_id:
        example: user-uuid-string
        type: string
    type: object
...
  dto.ResponseFormat:
    properties:
      data: {}
      error_code:
        type: integer
      error_message:
        type: string
      message:
        type: string
    type: object
  dto.SignupRequest:
    properties:
      check_password:
        type: string
      email:
        type: string
      nickname:
        maxLength: 20
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - check_password
    - email
    - nickname
    - password
    type: object
...
  dto.UpdateNicknameRequest:
    properties:
      nickname:
        maxLength: 20
        minLength: 2
        type: string
    required:
    - nickname
    type: object
  dto.UpdatePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
  dto.UserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: user-uuid-string
        type: string
    type: object
host: 10.30.8.25:8888
info:
  contact: {}
  description: keeplo
  title: keeplo API
  version: "0.1"
paths:
  /auth/duplicate:
    get:
      consumes:
      - application/json
      description: 입력한 이메일이 이미 사용 중인지 확인합니다.
      parameters:
      - description: 중복 확인할 이메일
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/dto.DuplicateEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ResponseFormat'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 이메일 중복 확인
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 이메일과 비밀번호로 로그인을 수행합니다.
      parameters:
      - description: 로그인 요청 정보
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ResponseFormat'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 로그인
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: 현재 로그인한 사용자의 정보를 반환합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ResponseFormat'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 내 정보 조회
      tags:
      - auth
...
  /auth/me/nickname:
    put:
      consumes:
      - application/json
      description: 사용자의 닉네임을 수정합니다.
      parameters:
      - description: 변경할 닉네임
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateNicknameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 닉네임 변경
      tags:
      - auth
  /auth/me/password:
    put:
      consumes:
      - application/json
      description: 기존 비밀번호를 새로운 비밀번호로 변경합니다.
      parameters:
      - description: 비밀번호 변경 정보
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 비밀번호 변경
      tags:
      - auth
  /auth/me/resign:
    delete:
      consumes:
      - application/json
      description: 현재 로그인한 사용자를 탈퇴 처리합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 회원 탈퇴
      tags:
      - auth
  /auth/password:
    post:
      consumes:
      - application/json
      description: 입력한 비밀번호가 현재 비밀번호와 일치하는지 확인합니다.
      parameters:
      - description: 확인할 비밀번호
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/dto.CheckPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 비밀번호 확인
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: 신규 사용자를 등록합니다.
      parameters:
      - description: 회원가입 요청 정보
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 회원 가입
      tags:
      - auth
...
schemes:
- http
swagger: "2.0"